+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - feat:
   - MT_MessageError and MT_MessageInfo for handle MetaTrader messages.
   - "get_error_messages" and "get_info_messages" new methods from MT_Client.
+- docs:
+  - README completed.
+  - LICENSE added.
 
 ## ** V0.14.0  (08/04/2024) **
 - refactor: